### 2024-06-20 #6907 @editorialbot
[PRE REVIEW]: snSMART: An R Package for Small Sample, Sequential, Multiple Assignment, Randomized Trial Data Analysis
**Submitting author:** @sidiwang (<a href="http://orcid.org/0000-0003-4838-0842">Sidi Wang</a>)
**Repository:** https://github.com/sidiwang/snSMART
**Branch with paper.md** (empty if default branch): JOSS
**Version:** v0.2.3
