### 2024-06-20 #6907 @crvernon
:wave:  @sidiwang - thanks for your submission to JOSS.  All of our the editors who would be able to edit this submission are currently occupied.  I am going to waitlist this submission until one becomes available.

In the meantime, I notice you specify a license in your package but there is no LICENSE file in the root of the repository.  Please add the appropriate open source license to your repository.  Also, at first glance, I don't see any tests accompanying your package.  This will come up if your submission makes it into full review so please try to take care of that ahead of time.  There are also many DOI in your references not accounted for (see report above).  Lastly, please reduce the length of the paper down to around 1000 words maximum.

Thanks!
