+++ paper.bib
@@ -9,7 +9,7 @@
   pages={105989},
   year={2020},
   publisher={Elsevier},
-  URL={https://doi.org/10.1016/j.cct.2020.105989}
+  doi={10.1016/j.cct.2020.105989}
 }
 
 @article{chao2020bayesian,
@@ -21,7 +21,7 @@
   pages={663--680},
   year={2020},
   publisher={Wiley Online Library},
-  URL={https://doi.org/10.1111/rssc.1246}
+  doi={10.1111/rssc.1246}
 }
 
 
@@ -35,7 +35,7 @@
   pages={3723--3732},
   year={2018},
   publisher={Wiley Online Library},
-  URL={https://doi.org/10.1002/sim.7900}
+  doi={10.1002/sim.7900}
 }
 
 
@@ -48,7 +48,7 @@
   pages={1109--1120},
   year={2020},
   publisher={Taylor \& Francis},
-  URL={https://doi.org/10.1080/10543406.2020.1815032}  
+  doi={10.1080/10543406.2020.1815032}  
 }
 
 
@@ -61,7 +61,7 @@
   pages={963--977},
   year={2021},
   publisher={Wiley Online Library},
-  URL={https://doi.org/10.1002/sim.8813}
+  doi={10.1002/sim.8813}
 }
 
 
@@ -74,7 +74,7 @@
   pages={312--326},
   year={2021},
   publisher={Wiley Online Library},
-  URL={https://doi.org/10.1002/sim.8776}
+  doi={10.1002/sim.8776}
 }
 
 
@@ -87,7 +87,7 @@
   pages={57--61},
   year={2002},
   publisher={Elsevier},
-  URL={https://doi.org/10.1016/S0140-6736(02)07283-5}
+  doi={10.1016/S0140-6736(02)07283-5}
 }
   
   
@@ -100,7 +100,7 @@
   pages={1--9},
   year={2020},
   publisher={BioMed Central},
-  URL={https://doi.org/10.1186/s13063-020-04285-3}
+  doi={10.1186/s13063-020-04285-3}
 }
 
 
@@ -113,7 +113,7 @@
   pages={433--439},
   year={1988},
   publisher={Wiley Online Library},
-  URL={https://doi.org/10.2307/2348770}
+  doi={10.2307/2348770}
 }
 
 
@@ -126,7 +126,7 @@
   pages={29--38},
   year={2000},
   publisher={Wiley Online Library},
-  URL={https://doi.org/10.1111/1467-985X.00154}
+  doi={10.1111/1467-985X.00154}
 }
   
 @article{murphy2005experimental,
@@ -138,7 +138,7 @@
   pages={1455--1481},
   year={2005},
   publisher={Wiley Online Library},
-  URL={https://doi.org/10.1002/sim.2022}
+  doi={10.1002/sim.2022}
 }
   
   
@@ -150,7 +150,7 @@
     journal = {Biostatistics},
     volume = {13},
     pages = {142-152},
-    URL={https://doi.org/10.1093/biostatistics/kxr016}
+    doi={10.1093/biostatistics/kxr016}
   }
 
 @Manual{geepackto,
@@ -158,7 +158,7 @@
     author = {Søren Højsgaard and Ulrich Halekoh and Jun Yan},
     year = {2022},
     note = {R package version 1.3.9},
-    url = {https://CRAN.R-project.org/package=geepack},
+    doi = {10.32614/CRAN.package.geepack},
   }
   
 @Article{gee1,
@@ -168,7 +168,7 @@
     journal = {Journal of Statistical Software},
     volume = {15/2},
     pages = {1--11},
-    URL={https://doi.org/10.18637/jss.v015.i02}
+    doi={10.18637/jss.v015.i02}
   }
 
[…]
