### 2024-06-26 bfef1d6
Update paper.md
