+++ paper.md
@@ -36,17 +36,15 @@ Small sample, sequential, multiple assignment, randomized trials (snSMARTs) are 
 
 # Statement of need
 
-The design and methods of snSMARTs generally apply to any disorder or disease that affects a small group of people and remains stable over the duration of the trial. A two-stage snSMART design first randomizes all participants to one of the first-stage treatments, then conducts second stage randomization based on its first stage treatment outcome. The goal of an snSMART is to identify the superior first stage treatment using data from both stages. An snSMART design repeats the same first stage treatments in the second stage of the trial, and may allow responders to the first stage treatment to stay on that same treatment in the second stage. Thus, the snSMART design allows for the estimation of first stage treatment effects by utilizing data from both stages, which is helpful for achieving more precise estimates.
-
-In the past few years, substantial progress has been made regarding the development of statistical methods for analyzing trial data from the following snSMART designs: an snSMART with three active treatments [@wei2018bayesian; @wei2020sample; @chao2020dynamic], a group sequential snSMART with three active treatments [@chao2020bayesian], an snSMART with placebo and two dose levels of one treatment [@fang2021bayesian], and an snSMART with continuous outcomes [@hartman2021design]. However, there is a lack of statistical software to disseminate the methods. Our R package `snSMART` fills this gap by providing functions for calculating the required sample size for an snSMART with three active treatments, and analyzing trial data using both Bayesian and frequentist approaches. To the extent of our knowledge, there are no existing R packages that provide similar functions. 
-
+The design and methods of snSMARTs generally apply to any disorder or disease that affects a small group of people and remains stable over the duration of the trial. Recent progress includes methods for snSMARTs with three active treatments [@wei2018bayesian; @wei2020sample; @chao2020dynamic], group sequential designs [@chao2020bayesian], placebo and two dose levels [@fang2021bayesian], and continuous outcomes [@hartman2021design]. However, there is a lack of software for these methods. The snSMART R package fills this gap by providing sample size calculations and trial data analysis using Bayesian and frequentist approaches. To our knowledge, no other R packages offer similar functions.
 
 ## snSMART comparing two dose levels with placebo
+
 Here we explain one of the snSMART designs in detail. This snSMART design investigates the response rate of one experimental treatment at low and high doses compared with placebo [@fang2021bayesian]. In such an snSMART design (Figure \ref{fig:snSMART-dose}), participants are equally assigned to either receive placebo, low dose, or high dose in the first stage. Participants receive their initial treatment for a pre-specified amount of time until the measurement of their responses at the end of stage 1. In the second stage, all participants who received placebo or low dose in the first stage are re-randomized to either low or high dose regardless of their first stage response. Participants who responded to high dose are re-randomized between low and high dose, whereas those who did not respond to high dose receive high dose again in the second stage. The main goal of this snSMART is to estimate first stage low and high dose treatment response rates and compare them to placebo by modeling the pooled data from two stages.
 
 ![\label{fig:snSMART-dose}](dose_snSMART.png)
 
[…]
