### 2024-06-26 24afed5
Update paper.md
