+++ paper.md
@@ -32,19 +32,19 @@ bibliography: paper.bib
 
 # Summary
 
-Small sample, sequential, multiple assignment, randomized trials (snSMARTs) are multistage trials with the primary goal of estimating treatment effects. Participants are first randomized to one of the first stage treatments, then participants may be re-randomized in the second stage depending on their outcome or response to first stage treatment. In an snSMART, treatment effects in the first stage are estimated using data from both stages, which results in more precise estimates. The design of the snSMART may also help to improve participant recruitment and retention over standard rare disease designs. In the past few years, substantial progress has been made regarding the development of statistical methods for snSMART designs. To better facilitate the application of these statistical methods, we introduce the R package snSMART in this paper, which can be used to analyze data from an existing snSMART design or aid in the design and sample size of an snSMART.
+Small sample, sequential, multiple assignment, randomized trials (snSMARTs) are multistage trials designed to estimate first stage treatment effects. By using data from all stages, snSMARTs provide more precise estimates of these effects. Additionally, the design may enhance participant recruitment and retention compared to standard rare disease trials. To support the application of snSMART statistical methods, we introduce the R package `snSMART`.
 
 # Statement of need
 
-The design and methods of snSMARTs generally apply to any disorder or disease that affects a small group of people and remains stable over the duration of the trial. Recent progress includes methods for snSMARTs with three active treatments [@wei2018bayesian; @wei2020sample; @chao2020dynamic], group sequential designs [@chao2020bayesian], placebo and two dose levels [@fang2021bayesian], and continuous outcomes [@hartman2021design]. However, there is a lack of software for these methods. The snSMART R package fills this gap by providing sample size calculations and trial data analysis using Bayesian and frequentist approaches. To our knowledge, no other R packages offer similar functions.
+The design and methods of snSMARTs are applicable to any disorder or disease that affects a small group and remains stable over the trial duration. Recent advances include methods for snSMARTs with three active treatments [@wei2018bayesian; @wei2020sample; @chao2020dynamic], group sequential designs [@chao2020bayesian], placebo with two dose levels [@fang2021bayesian], and continuous outcomes [@hartman2021design]. Despite these developments, there is a lack of software to implement these methods. The `snSMART` R package addresses this need by providing sample size calculations and trial data analysis using both Bayesian and frequentist approaches. To our knowledge, no other R packages offer similar functionalities.
 
 ## snSMART comparing two dose levels with placebo
 
[…]
