+++ R/JSRM_binary.R
@@ -260,9 +260,10 @@ summary.LPJSM_binary <- function(object, ...) {
     DTRest <- t(rbind(object$pi_DTR_hat, object$pi_DTR_se))
     colnames(DTRest) <- c("Estimate", "Std. Error")
     rownames(DTRest) <- c("rep_AB", "rep_AC", "rep_BA", "rep_BC", "rep_CA", "rep_CB")
+    obj <- list(GEEresult = summary(object$GEE_output), trteff = trteff, DTRest = DTRest)
   }
 
-  obj <- list(GEEresult = summary(object$GEE_output), trteff = trteff, DTRest = DTRest)
+  obj <- list(GEEresult = summary(object$GEE_output), trteff = trteff)
   class(obj) <- "summary.LPJSM_binary"
   obj
 }
