+++ tests/testthat.R
@@ -1,0 +1,393 @@
+# This file is part of the standard setup for testthat.
+# It is recommended that you do not modify it.
+#
+# Where should you do additional test configuration?
+# Learn more about the roles of various files in:
+# * https://r-pkgs.org/testing-design.html#sec-tests-files-overview
+# * https://testthat.r-lib.org/articles/special-files.html
+
+library(testthat)
+library(snSMART)
+
+test_that("BJSM_binary 1", {
+  mydata <- data_binary
+
+  BJSM_result <- BJSM_binary(
+    data = mydata, prior_dist = c("beta", "beta", "pareto"),
+    pi_prior = c(0.4, 1.6, 0.4, 1.6, 0.4, 1.6), beta_prior = c(1.6, 0.4, 3, 1),
+    n_MCMC_chain = 1, n.adapt = 1000, MCMC_SAMPLE = 2000, ci = 0.95,
+    six = TRUE, DTR = TRUE, verbose = FALSE
+  )
+
+  result = c(0.3889, 0.4371, 0.5793)
+  names(result) = c("pi_A", "pi_B", "pi_C")
+
+  expect_equal(BJSM_result$pi_hat_bjsm, result, tolerance = 1e-1)
+})
+
+test_that("BJSM_binary 2", {
+  mydata <- data_binary
+
+  BJSM_result <- BJSM_binary(
+    data = mydata, prior_dist = c("beta", "beta", "pareto"),
+    pi_prior = c(0.4, 1.6, 0.4, 1.6, 0.4, 1.6), beta_prior = c(1.6, 0.4, 3, 1),
+    n_MCMC_chain = 1, n.adapt = 1000, MCMC_SAMPLE = 2000, ci = 0.95,
+    six = TRUE, DTR = FALSE, verbose = FALSE
+  )
+
+  result = c(0.3889, 0.4371, 0.5793)
+  names(result) = c("pi_A", "pi_B", "pi_C")
+
+  expect_equal(BJSM_result$pi_hat_bjsm, result, tolerance = 1e-1)
+})
+
+test_that("BJSM_binary 3", {
+  mydata <- data_binary
+
+  BJSM_result <- BJSM_binary(
+    data = mydata, prior_dist = c("beta", "beta", "pareto"),
+    pi_prior = c(0.4, 1.6, 0.4, 1.6, 0.4, 1.6), beta_prior = c(1.6, 0.4, 3, 1),
+    n_MCMC_chain = 1, n.adapt = 1000, MCMC_SAMPLE = 2000, ci = 0.95,
+    six = TRUE, DTR = TRUE, verbose = TRUE
+  )
+
+  result = c(0.3889, 0.4371, 0.5793)
+  names(result) = c("pi_A", "pi_B", "pi_C")
+
+  expect_equal(BJSM_result$pi_hat_bjsm, result, tolerance = 1e-1)
+})
+
+test_that("BJSM_binary 4", {
+  mydata <- data_binary
+
+  BJSM_result <- BJSM_binary(
+    data = mydata, prior_dist = c("beta", "beta", "pareto"),
+    pi_prior = c(0.4, 1.6, 0.4, 1.6, 0.4, 1.6), beta_prior = c(1.6, 0.4, 3, 1),
+    n_MCMC_chain = 1, n.adapt = 1000, MCMC_SAMPLE = 2000, ci = 0.95,
+    six = TRUE, DTR = FALSE, verbose = FALSE
+  )
+
+  result = c(0.3889, 0.4371, 0.5793)
+  names(result) = c("pi_A", "pi_B", "pi_C")
+
[…]
