+++ tests/testthat.R
@@ -19,6 +19,10 @@ test_that("BJSM_binary 1", {
     six = TRUE, DTR = TRUE, verbose = FALSE
   )
 
+  summary(BJSM_result)
+  print(summary(BJSM_result))
+  print(BJSM_result)
+
   result = c(0.3889, 0.4371, 0.5793)
   names(result) = c("pi_A", "pi_B", "pi_C")
 
@@ -34,6 +38,10 @@ test_that("BJSM_binary 2", {
     n_MCMC_chain = 1, n.adapt = 1000, MCMC_SAMPLE = 2000, ci = 0.95,
     six = TRUE, DTR = FALSE, verbose = FALSE
   )
+
+  summary(BJSM_result)
+  print(summary(BJSM_result))
+  print(BJSM_result)
 
   result = c(0.3889, 0.4371, 0.5793)
   names(result) = c("pi_A", "pi_B", "pi_C")
@@ -51,6 +59,10 @@ test_that("BJSM_binary 3", {
     six = TRUE, DTR = TRUE, verbose = TRUE
   )
 
+  summary(BJSM_result)
+  print(summary(BJSM_result))
+  print(BJSM_result)
+
   result = c(0.3889, 0.4371, 0.5793)
   names(result) = c("pi_A", "pi_B", "pi_C")
 
@@ -66,6 +78,10 @@ test_that("BJSM_binary 4", {
     n_MCMC_chain = 1, n.adapt = 1000, MCMC_SAMPLE = 2000, ci = 0.95,
     six = TRUE, DTR = FALSE, verbose = FALSE
   )
+
+  summary(BJSM_result)
+  print(summary(BJSM_result))
+  print(BJSM_result)
 
   result = c(0.3889, 0.4371, 0.5793)
   names(result) = c("pi_A", "pi_B", "pi_C")
@@ -83,6 +99,10 @@ test_that("BJSM_binary 5", {
     six = FALSE, DTR = FALSE, verbose = FALSE
   )
 
+  summary(BJSM_result2)
+  print(BJSM_result2)
+  print(summary(BJSM_result2))
+
   result = c(0.3993, 0.4250, 0.5411)
   names(result) = c("pi_A", "pi_B", "pi_C")
 
@@ -99,6 +119,10 @@ test_that("BJSM_binary 6", {
     six = FALSE, DTR = TRUE, verbose = FALSE
   )
 
+  summary(BJSM_result2)
+  print(BJSM_result2)
+  print(summary(BJSM_result2))
+
   result = c(0.3993, 0.4250, 0.5411)
   names(result) = c("pi_A", "pi_B", "pi_C")
 
@@ -114,6 +138,10 @@ test_that("BJSM_binary 7", {
     n_MCMC_chain = 1, n.adapt = 10000, MCMC_SAMPLE = 60000, ci = 0.95,
     six = FALSE, DTR = TRUE, verbose = TRUE
   )
+
+  summary(BJSM_result2)
+  print(BJSM_result2)
+  print(summary(BJSM_result2))
 
   result = c(0.3993, 0.4250, 0.5411)
   names(result) = c("pi_A", "pi_B", "pi_C")
@@ -132,6 +160,10 @@ test_that("BJSM_binary 7", {
     six = FALSE, DTR = FALSE, verbose = TRUE
   )
 
+  summary(BJSM_result2)
+  print(BJSM_result2)
+  print(summary(BJSM_result2))
+
   result = c(0.3993, 0.4250, 0.5411)
   names(result) = c("pi_A", "pi_B", "pi_C")
 
@@ -148,6 +180,10 @@ test_that("BJSM_binary 8", {
[…]
