+++ codecov.yml
@@ -12,3 +12,6 @@ coverage:
         target: auto
         threshold: 1%
         informational: true
+codecov:
+  upload:
+    file: coverage.xml
