+++ codecov.yml
@@ -12,6 +12,3 @@ coverage:
         target: auto
         threshold: 1%
         informational: true
-codecov:
-  upload:
-    file: coverage.xml
