+++ codecov.yml
@@ -12,15 +12,3 @@ coverage:
         target: auto
         threshold: 1%
         informational: true
-
-comment: false
-
-language: R
-
-sudo: false
-
-cache: packages
-
-after_success:
-
-- Rscript -e 'covr::codecov()'
