+++ .github/workflows/R-CMD-check.yaml
@@ -2,9 +2,9 @@
 # Need help debugging build failures? Start at https://github.com/r-lib/actions#where-to-find-help
 on:
   push:
-    branches: [main, master]
+    branches: [main, cran, master]
   pull_request:
-    branches: [main, master]
+    branches: [main, cran, master]
 
 name: R-CMD-check
 
