### 2024-07-06 #6907 @osorensen
👋 @gasparl, @aghaynes, @rsh52, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
