### 2024-07-09 #6907 @rsh52
Thanks @osorensen!

My colleague and fellow co-author on our JOSS paper, @ezraporter, might be a better choice for this given his stats background.
