### 2024-07-10 #6907 @ezraporter
@osorensen yes, I'd be happy to!
