### 2024-07-10 #6971 @editorialbot
[REVIEW]: snSMART: An R Package for Small Sample, Sequential, Multiple Assignment, Randomized Trial Data Analysis
**Submitting author:** @sidiwang (<a href="http://orcid.org/0000-0003-4838-0842">Sidi Wang</a>)
**Repository:** https://github.com/sidiwang/snSMART
**Branch with paper.md** (empty if default branch): JOSS
**Version:** 0.2.4
