### 2024-07-16 sidiwang/snSMART#23 @sidiwang
Thank you very much for your feedback!

Regarding the authorship, I will discuss this with all the authors and get back to you.

The datasets are synthetic; I will add this to the descriptions.

I will make changes according to all the other points you mentioned asap and will let you know when they are done. Just out of curiosity, is there a deadline for this?

Thanks again for reviewing the paper.
