### 2024-07-17 sidiwang/snSMART#23 @aghaynes
You're welcome. If you have any questions, let me know 😄

Re the review, thats a little unclear to me too, although the editorbot said "please make sure to complete your review within six weeks" in the review issue. Whether that includes your response/changes I don't know.
