+++ R/JSRM_binary.R
@@ -30,15 +30,13 @@
 #' summary(LPJSM_result)
 #'
 #' @references
-#' Wei, B., Braun, T.M., Tamura, R.N. and Kidwell, K.M., 2018. A Bayesian analysis
-#' of small n sequential multiple assignment randomized trials (snSMARTs).
-#' Statistics in medicine, 37(26), pp.3723-3732.
+#' Wei, B., Braun, T.M., Tamura, R.N. and Kidwell, K.M., 2018. A Bayesian analysis of small n sequential multiple assignment randomized trials (snSMARTs).
+#' Statistics in medicine, 37(26), pp.3723-3732. URL: [https://doi.org/10.1002/sim.7900](https://doi.org/10.1002/sim.7900)
 #'
-#' Chao, Y.C., Trachtman, H., Gipson, D.S., Spino, C., Braun, T.M. and Kidwell, K.M., 2020. Dynamic treatment regimens in small n, sequential, multiple assignment,
-#' randomized trials: An application in focal segmental glomerulosclerosis. Contemporary clinical trials, 92, p.105989.
+#' Chao, Y.C., Trachtman, H., Gipson, D.S., Spino, C., Braun, T.M. and Kidwell, K.M., 2020. Dynamic treatment regimens in small n, sequential, multiple assignment, randomized trials: An application in focal segmental glomerulosclerosis. Contemporary clinical trials, 92, p.105989. URL: [https://doi.org/10.1016/j.cct.2020.105989](https://doi.org/10.1016/j.cct.2020.105989)
 #'
 #' Fang, F., Hochstedler, K.A., Tamura, R.N., Braun, T.M. and Kidwell, K.M., 2021. Bayesian methods to compare dose levels with placebo in a small n,
-#' sequential, multiple assignment, randomized trial. Statistics in Medicine, 40(4), pp.963-977.
+#' sequential, multiple assignment, randomized trial. Statistics in Medicine, 40(4), pp.963-977. URL: [https://doi.org/10.1002/sim.8813](https://doi.org/10.1002/sim.8813)
 #'
 #' @seealso
 #' \code{\link{BJSM_binary}} \cr
