+++ man/groupseqDATA_look1.Rd
@@ -21,7 +21,7 @@ This data frame contains the following columns:
 groupseqDATA_look1
 }
 \description{
-sample dataset of group sequential trial design snSMART, can be used for interim analysis
+sample synthetic dataset of group sequential trial design snSMART, can be used for interim analysis
 }
 \examples{
 mydata <- groupseqDATA_look1
