### 2024-08-09 95c4426
Delete paper.md
