+++ paper.md
@@ -1,171 +1,0 @@
----
-title: "snSMART: An R Package for Small Sample, Sequential, Multiple Assignment, Randomized Trial Data Analysis"
-tags:
-  - rare disease
-  - clinical trial
-  - snSMART
-  - Bayesian
-authors:  
-  - name: Sidi Wang
-    orcid: 0000-0003-4838-0842
-    affiliation: 1
-  - name: Fang Fang
-    orcid: 0000-0002-7089-3591
-    affiliation: 1
-  - name: Roy Tamura
-    affiliation: 2
-  - name: Thomas Braun
-    orcid: 0000-0002-7113-6998
-    affiliation: 1
-  - name: Kelley M Kidwell
-    orcid: 0000-0002-1717-4483
-    affiliation: 1
-affiliations:
-  - name: University of Michigan
-    index: 1
-  - name: University of South Florida
-    index: 2
-date: 22 May 2024
-bibliography: paper.bib
-
----
-
-# Summary
-
-Small sample, sequential, multiple assignment, randomized trials (snSMARTs) are multistage trials designed to estimate first stage treatment effects. By using data from all stages, snSMARTs provide more precise estimates of these effects. Additionally, the design may enhance participant recruitment and retention compared to standard rare disease trials. To support the application of snSMART statistical methods, we introduce the R package `snSMART`.
-
-# Statement of need
-
-The design and methods of snSMARTs are applicable to any disorder or disease that affects a small group and remains stable over the trial duration. Recent advances include methods for snSMARTs with three active treatments [@wei2018bayesian; @wei2020sample; @chao2020dynamic], group sequential designs [@chao2020bayesian], placebo with two dose levels [@fang2021bayesian], and continuous outcomes [@hartman2021design]. Despite these developments, there is a lack of software to implement these methods. The `snSMART` R package addresses this need by providing sample size calculations and trial data analysis using both Bayesian and frequentist approaches. To our knowledge, no other R packages offer similar functionalities.
-
-## snSMART comparing two dose levels with placebo
-
-This section details one of the snSMART designs, which investigates the response rate of an experimental treatment at low and high doses compared to placebo [@fang2021bayesian]. In this design (Figure \ref{fig:snSMART-dose}), participants are equally assigned to receive either placebo, low dose, or high dose in the first stage. They continue their initial treatment for a pre-specified duration until their responses are measured at the end of stage 1. In the second stage, all participants who initially received placebo or low dose are re-randomized to either low or high dose, regardless of their first stage response. Participants who responded to the high dose are re-randomized between low and high doses, while non-responders to the high dose continue with the high dose in the second stage. The main goal of this snSMART is to estimate and compare first stage response rates for low and high doses to placebo by modeling the pooled data from both stages.
-
-![Study design of an snSMART with two dose levels and a placebo. In stage 1, participants are randomized (R) to treatment P (placebo), L (low dose), or H (high dose) with equal probability. At time t, response to stage 1 treatment is assessed. Non-responders to high dose stay on the same treatment in stage 2, while all the other participants are equally re-randomized to either low or high dose in stage 2. Interest is in the first stage response rate of placebo, low and high doses.\label{fig:snSMART-dose}](dose_snSMART.png)
-
[…]
