### 2024-08-09 1b68b16
Delete paper.bib
