### 2024-08-09 sidiwang/snSMART#23 @sidiwang
Hi @aghaynes,

Thank you for the feedback! I have moved the contributing.md file, added it to the .Rbuildignore file, and added row names to the result discussions. I hope they look better now!

Thanks again for your review.

Regards,
Sidi
