### 2024-08-12 sidiwang/snSMART#23 @aghaynes
Hi @sidiwang
I just noticed that you mention a contributor code of conduct in your contributing.md, but the file doesnt seem to be there... it would also be fine to put everything in one file in my opinion.
