+++ paper.bib
@@ -1,0 +1,229 @@
+% put bibtex entries here!
+
+
+@article{chao2020dynamic,
+  title={Dynamic Treatment Regimens in Small N, Sequential, Multiple Assignment, Randomized Trials: An Application in Focal Segmental Glomerulosclerosis},
+  author={Chao, Yan-Cheng and Trachtman, Howard and Gipson, Debbie S and Spino, Cathie and Braun, Thomas M and Kidwell, Kelley M},
+  journal={Contemporary Clinical Trials},
+  volume={92},
+  pages={105989},
+  year={2020},
+  publisher={Elsevier},
+  doi={10.1016/j.cct.2020.105989}
+}
+
+@article{chao2020bayesian,
+  title={A {Bayesian} Group Sequential Small N Sequential Multiple-Assignment Randomized Trial},
+  author={Chao, Yan-Cheng and Braun, Thomas M and Tamura, Roy N and Kidwell, Kelley M},
+  journal={Journal of the Royal Statistical Society C},
+  volume={69},
+  number={3},
+  pages={663--680},
+  year={2020},
+  publisher={Wiley Online Library},
+  doi={10.1111/rssc.1246}
+}
+
+
+
+@article{wei2018bayesian,
+  title={A {Bayesian} Analysis of Small N Sequential Multiple Assignment Randomized Trials ({snSMARTs})},
+  author={Wei, Boxian and Braun, Thomas M and Tamura, Roy N and Kidwell, Kelley M},
+  journal={Statistics in Medicine},
+  volume={37},
+  number={26},
+  pages={3723--3732},
+  year={2018},
+  publisher={Wiley Online Library},
+  doi={10.1002/sim.7900}
+}
+
+
+@article{wei2020sample,
+  title={Sample Size Determination for {Bayesian} Analysis of Small N Sequential, Multiple Assignment, Randomized Trials ({snSMARTs}) with Three Agents},
+  author={Wei, Boxian and Braun, Thomas M and Tamura, Roy N and Kidwell, Kelley},
+  journal={Journal of Biopharmaceutical Statistics},
+  volume={30},
+  number={6},
+  pages={1109--1120},
+  year={2020},
+  publisher={Taylor \& Francis},
+  doi={10.1080/10543406.2020.1815032}  
+}
+
+
+@article{fang2021bayesian,
+  title={Bayesian Methods to Compare Dose Levels with Placebo in a Small N, Sequential, Multiple Assignment, Randomized Trial},
+  author={Fang, Fang and Hochstedler, Kimberly A and Tamura, Roy N and Braun, Thomas M and Kidwell, Kelley M},
+  journal={Statistics in Medicine},
+  volume={40},
+  number={4},
+  pages={963--977},
+  year={2021},
+  publisher={Wiley Online Library},
+  doi={10.1002/sim.8813}
+}
+
+
+@article{hartman2021design,
+  title={Design and Analysis Considerations for Utilizing a Mapping Function in a Small Sample, Sequential, Multiple Assignment, Randomized Trials with Continuous Outcomes},
+  author={Hartman, Holly and Tamura, Roy N and Schipper, Matthew J and Kidwell, Kelley M},
+  journal={Statistics in Medicine},
+  volume={40},
+  number={2},
+  pages={312--326},
+  year={2021},
+  publisher={Wiley Online Library},
+  doi={10.1002/sim.8776}
+}
+
+
+@article{grimes2002overview,
+  title={An Overview of Clinical Research: the Lay of the Land},
+  author={Grimes, David A and Schulz, Kenneth F},
+  journal={The Lancet},
+  volume={359},
+  number={9300},
+  pages={57--61},
+  year={2002},
+  publisher={Elsevier},
[…]
