### 2024-08-13 sidiwang/snSMART#23 @sidiwang
Hi @aghaynes,

Thank you! I forgot to include that file. Just added it. Please check the updated JOSS branch. :)

Thanks,
Sidi
