+++ README.md
@@ -36,7 +36,7 @@ across the two stages of the snSMART design to more precisely estimate
 the effect of the treatments given in the first stage.
 
 ## Installation
-Before using this package, please install the \href{https://sourceforge.net/projects/mcmc-jags/}{JAGS library} on your device.
+Before using this package, please install the [JAGS library](https://sourceforge.net/projects/mcmc-jags/) on your device.
 
 You can install the snSMART package from CRAN:
 
