### 2024-08-18 7e3f123
edite n_MCMC_chain default
