### 2024-08-18 1bc967b
updated the paper and LPJSM_binary documentation
