+++ README.md
@@ -36,7 +36,9 @@ across the two stages of the snSMART design to more precisely estimate
 the effect of the treatments given in the first stage.
 
 ## Installation
-Before using this package, please install the [JAGS library](https://sourceforge.net/projects/mcmc-jags/) on your device.
+
+Before using this package, please install the [JAGS
+library](https://sourceforge.net/projects/mcmc-jags/) on your device.
 
 You can install the snSMART package from CRAN:
 
@@ -46,7 +48,7 @@ install.packages("snSMART", repos = "http://cran.us.r-project.org")
 
     ## 
     ## The downloaded binary packages are in
-    ##  /var/folders/5_/sp2r7r_s5snf4hq634xmk_0r0000gn/T//RtmpNRlYCm/downloaded_packages
+    ##  /var/folders/5_/sp2r7r_s5snf4hq634xmk_0r0000gn/T//RtmpUxY3P4/downloaded_packages
 
 ``` r
 library(snSMART)
@@ -108,33 +110,33 @@ summary(BJSM_result)
     ## 
     ## Treatment Effects Estimate:
     ##       Estimate Std. Error C.I.    CI low   CI high
-    ## trtA 0.4046939 0.10427533 0.95 0.2103586 0.6583262
-    ## trtB 0.3900296 0.07865828 0.95 0.2501780 0.5374523
-    ## trtC 0.6180690 0.08709180 0.95 0.4925770 0.7772426
+    ## trtA 0.3733798  0.1532599 0.95 0.1879501 0.7281596
+    ## trtB 0.4185251  0.1646755 0.95 0.1377496 0.7658878
+    ## trtC 0.6694094  0.1044533 0.95 0.4533836 0.8364925
     ## 
     ## Differences between Treatments:
     ##           Estimate Std.Error C.I.     CI low     CI high
-    ## diffAB  0.01466433 0.1287876 0.95 -0.2202233  0.32554853
-    ## diffBC -0.22803936 0.1271915 0.95 -0.5270646 -0.06082518
-    ## diffAC -0.21337503 0.1493975 0.95 -0.4840569  0.04220399
+    ## diffAB -0.04514528 0.2534023 0.95 -0.5669267 0.377003340
+    ## diffBC -0.25088433 0.1648132 0.95 -0.5241105 0.093336459
+    ## diffAC -0.29602961 0.1676356 0.95 -0.5361779 0.002146219
     ## 
     ## Linkage Parameter Estimate:
     ##         Estimate Std. Error C.I.    CI low   CI high
-    ## beta0A 0.9595907 0.04719401 0.95 0.8544802 0.9998446
-    ## beta0B 0.9575869 0.05057664 0.95 0.8615737 0.9997691
-    ## beta0C 0.6930520 0.24176704 0.95 0.3444717 0.9917792
-    ## beta1A 1.5171964 0.38056743 0.95 1.0808245 2.3459093
-    ## beta1B 1.7896323 0.56996358 0.95 1.1208255 2.9052050
-    ## beta1C 1.3295262 0.23142426 0.95 1.0005293 1.8001938
+    ## beta0A 0.9109306 0.09415977 0.95 0.7450933 0.9991757
+    ## beta0B 0.9474557 0.05466744 0.95 0.8546296 0.9998542
+    ## beta0C 0.6844260 0.25597435 0.95 0.1968036 0.9921257
+    ## beta1A 1.3743002 0.40896514 0.95 1.0046594 2.9088154
+    ## beta1B 1.6303701 0.81305930 0.95 1.0023232 3.8184414
[…]
