### 2024-08-19 sidiwang/snSMART#30 @sidiwang
Hi @ezraporter,

Thank you very much for your valuable suggestions. I have reviewed them one by one and made edits according to every bullet point you listed under "Installation," "Documentation," and "Manuscript". Could you please take a look at the repository again and let me know if you feel further changes are needed? Regarding the "Additional suggestions not impacting the review," I will make those changes in the next few weeks.

Thanks,
Sidi
