### 2024-08-24 #6971 @osorensen
@sidiwang I've open to pull requests with small suggested changes to the paper. Please ping me here when they're fixed.
