### 2024-08-24 #6971 @sidiwang
Hi @osorensen,

Thank you for your suggestions! I have made the changes. Please take a look and let me know if further adjustments are needed.

Thanks,
Sidi
