### 2024-08-24 #6971 @osorensen
@sidiwang, thanks. It seems like the reference to Chao et al. (2020) has an error in the DOI. The correct DOI is 10.1111/rssc.12406. Could you please update?
