### 2024-08-24 #6971 @sidiwang
Hi @osorensen,

I just made a tagged release, the version is 0.2.4, and the DOI is 10.5281/zenodo.13370298

I hope everything was done correctly. Please let me know if there’s anything that needs adjustment.

Thanks,
Sidi
