### 2024-08-26 #6971 @osorensen
@openjournals/dev can you help us with the issue above?
