### 2024-08-29 #6971 @arfon
Looks like there's some issue with the generated JATS file: https://github.com/openjournals/joss-papers/blob/bcb1737a9ecc52b458bd1019dd24fa82f36d10e6/joss.06971/paper.jats/10.21105.joss.06971.jats

I think it's the table in the document. @sidiwang – could you please try updating your table to be strictly Markdown formatted one?
