### 2024-09-03 #6971 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
