### 2024-09-06 #6971 @samhforbes
Hi @sidiwang I noticed your software licence doesn't match the licence in your archive. Can you update to make sure both use a [OSI approved licence](https://opensource.org/licenses) and that it matches please?
