### 2024-09-06 #6971 @sidiwang
Hi @samhforbes I just updated the license in my archive. Hopefully I did it in the right way! Thank you!
